divided upstream mocks.h into different classes
