+++ test/mocks/upstream/basic_resource_limit.cc
@@ -1,0 +1,17 @@
+#include "basic_resource_limit.h"
+
+#include "gmock/gmock.h"
+#include "gtest/gtest.h"
+
+namespace Envoy {
+namespace Upstream {
+
+using ::testing::Return;
+MockBasicResourceLimit::MockBasicResourceLimit() {
+  ON_CALL(*this, canCreate()).WillByDefault(Return(true));
+}
+
+MockBasicResourceLimit::~MockBasicResourceLimit() = default;
+
+} // namespace Upstream
+} // namespace Envoy
