+++ test/mocks/upstream/cluster_info_factory.cc
@@ -1,0 +1,10 @@
+#include "cluster_info_factory.h"
+
+namespace Envoy {
+namespace Upstream {
+MockClusterInfoFactory::MockClusterInfoFactory() = default;
+
+MockClusterInfoFactory::~MockClusterInfoFactory() = default;
+
+} // namespace Upstream
+} // namespace Envoy
