+++ test/mocks/upstream/cluster_manager_factory.cc
@@ -1,0 +1,9 @@
+#include "cluster_manager_factory.h"
+
+namespace Envoy {
+namespace Upstream {
+MockClusterManagerFactory::MockClusterManagerFactory() = default;
+
+MockClusterManagerFactory::~MockClusterManagerFactory() = default;
+} // namespace Upstream
+} // namespace Envoy
