+++ test/mocks/upstream/cluster_priority_set.cc
@@ -1,0 +1,10 @@
+#include "cluster_priority_set.h"
+
+namespace Envoy {
+namespace Upstream {
+MockClusterMockPrioritySet::MockClusterMockPrioritySet() = default;
+
+MockClusterMockPrioritySet::~MockClusterMockPrioritySet() = default;
+
+} // namespace Upstream
+} // namespace Envoy
