+++ test/mocks/upstream/cluster_real_priority_set.cc
@@ -1,0 +1,9 @@
+#include "cluster_real_priority_set.h"
+
+namespace Envoy {
+namespace Upstream {
+MockClusterRealPrioritySet::MockClusterRealPrioritySet() = default;
+
+MockClusterRealPrioritySet::~MockClusterRealPrioritySet() = default;
+} // namespace Upstream
+} // namespace Envoy
