+++ test/mocks/upstream/cluster_update_callbacks.cc
@@ -1,0 +1,10 @@
+#include "cluster_update_callbacks.h"
+
+namespace Envoy {
+namespace Upstream {
+MockClusterUpdateCallbacks::MockClusterUpdateCallbacks() = default;
+
+MockClusterUpdateCallbacks::~MockClusterUpdateCallbacks() = default;
+
+} // namespace Upstream
+} // namespace Envoy
