+++ test/mocks/upstream/cluster_update_callbacks_handle.cc
@@ -1,0 +1,9 @@
+#include "cluster_update_callbacks_handle.h"
+
+namespace Envoy {
+namespace Upstream {
+MockClusterUpdateCallbacksHandle::MockClusterUpdateCallbacksHandle() = default;
+
+MockClusterUpdateCallbacksHandle::~MockClusterUpdateCallbacksHandle() = default;
+} // namespace Upstream
+} // namespace Envoy
