+++ test/mocks/upstream/load_balancer_context.h
@@ -1,3 +1,4 @@
+#pragma once
 #include "envoy/upstream/load_balancer.h"
 
 #include "gmock/gmock.h"
