+++ test/mocks/upstream/retry_host_predicate.cc
@@ -1,0 +1,10 @@
+#include "retry_host_predicate.h"
+
+namespace Envoy {
+namespace Upstream {
+MockRetryHostPredicate::MockRetryHostPredicate() = default;
+
+MockRetryHostPredicate::~MockRetryHostPredicate() = default;
+
+} // namespace Upstream
+} // namespace Envoy
