+++ test/mocks/upstream/retry_priority.cc
@@ -1,0 +1,7 @@
+#include "retry_priority.h"
+
+namespace Envoy {
+namespace Upstream {
+MockRetryPriority::~MockRetryPriority() = default;
+}
+} // namespace Envoy
