+++ test/mocks/upstream/thread_aware_load_balancer.cc
@@ -1,0 +1,10 @@
+#include "thread_aware_load_balancer.h"
+
+namespace Envoy {
+namespace Upstream {
+MockThreadAwareLoadBalancer::MockThreadAwareLoadBalancer() = default;
+
+MockThreadAwareLoadBalancer::~MockThreadAwareLoadBalancer() = default;
+
+} // namespace Upstream
+} // namespace Envoy
